+++ src/isValid/index.js
@@ -1,40 +1,30 @@
-import breakdown from '../breakdown/index';
-import callingCodes from '../callingCodes.js';
-import uglify from '../uglify/index';
+import breakdown from '../breakdown';
+import uglify from '../uglify';
 
-const validateCountryCode = (phone, country) => {
-  const { countryCode } = breakdown(phone, true);
-
-  return callingCodes[country.toLowerCase()].indexOf(countryCode) !== -1;
-};
+/* eslint-disable no-useless-escape */
 
 /**
  * @name isValid
- * @description Checks if the phone number is valid or not
- *
+ * @since v0.1.0
+ * @category Function
+ * @description
+ * Validates the base number, does not take the country code or extension into consideration for this validation
  * @param {String} phone The phone number to breakdown
  * @param {String} country The country name to validate with
  * @return {Boolean} Returns a boolean if the number provided is valid or not
- *
+ *phone && (/^\+?([0-9]{2})\)?[-. ]?([0-9]{4})[-. ]?([0-9]{4})$/).test(uglify(phone));
  * @example
- * const results = isValid('555-444-3333'); // => true
- * const results = isValid('1 + (555) 444 3333', 'usa'); // => true
- * const results = isValid('9885554443333', 'usa'); // => false
+ * isValid('555-444-3333'); // => true
  */
-const isValid = (phone, country) => {
-  let valid = false;
+export default phone => {
+  const uglyPhone = uglify(phone);
 
-  if (phone && (/^[0-9]{7,}$/).test(uglify(phone))) {
-    valid = true;
-  } else {
+  if (!phone || uglyPhone.length < 10) {
     return false;
   }
 
-  if (country && !validateCountryCode(phone, country)) {
-    valid = false;
-  }
+  const { areaCode, localCode, lineNumber } = breakdown('', uglyPhone);
 
-  return valid;
+  return phone && (/^\+?([0-9]{2})\)?[-. ]?([0-9]{4})[-. ]?([0-9]{4})$/)
+    .test(areaCode + localCode + lineNumber);
 };
-
-export default isValid;